+++ lib/packages/wikibase/data-model-services/tests/unit/Lookup/InMemoryEntityLookupTest.php
@@ -4,9 +4,14 @@ namespace Wikibase\DataModel\Services\Tests\Lookup;
 
 use InvalidArgumentException;
 use Wikibase\DataModel\Entity\ItemId;
+use Wikibase\DataModel\Entity\PropertyId;
 use Wikibase\DataModel\Services\Fixtures\FakeEntityDocument;
+use Wikibase\DataModel\Services\Fixtures\ItemFixtures;
+use Wikibase\DataModel\Services\Fixtures\PropertyFixtures;
 use Wikibase\DataModel\Services\Lookup\EntityLookupException;
 use Wikibase\DataModel\Services\Lookup\InMemoryEntityLookup;
+use Wikibase\DataModel\Services\Lookup\ItemLookupException;
+use Wikibase\DataModel\Services\Lookup\PropertyLookupException;
 
 /**
  * @covers \Wikibase\DataModel\Services\Lookup\InMemoryEntityLookup
@@ -85,6 +90,60 @@ class InMemoryEntityLookupTest extends \PHPUnit_Framework_TestCase {
 		new InMemoryEntityLookup(
 			new FakeEntityDocument()
 		);
+	}
+
+	public function testGivenKnownItem_getItemForIdReturnsIt() {
+		$item = ItemFixtures::newItem();
+
+		$lookup = new InMemoryEntityLookup( $item );
+
+		$this->assertEquals(
+			$item,
+			$lookup->getItemForId( $item->getId() )
+		);
+	}
+
+	public function testGivenKnownProperty_getPropertyForIdReturnsIt() {
+		$property = PropertyFixtures::newProperty();
+
+		$lookup = new InMemoryEntityLookup( $property );
+
+		$this->assertEquals(
+			$property,
+			$lookup->getPropertyForId( $property->getId() )
+		);
+	}
+
+	public function testWhenItemIsNotKnown_getItemForIdReturnsNull() {
+		$this->assertNull(
+			( new InMemoryEntityLookup() )->getItemForId( new ItemId( 'Q1' ) )
+		);
+	}
+
+	public function testWhenPropertyIsNotKnown_getPropertyForIdReturnsNull() {
+		$this->assertNull(
+			( new InMemoryEntityLookup() )->getPropertyForId( new PropertyId( 'P1' ) )
+		);
+	}
+
+	public function testGetItemForIdThrowsCorrectExceptionType() {
+		$id = new ItemId( 'Q1' );
+
+		$lookup = new InMemoryEntityLookup();
+		$lookup->addException( new EntityLookupException( $id ) );
+
+		$this->expectException( ItemLookupException::class );
+		$lookup->getItemForId( $id );
+	}
+
+	public function testGetPropertyForIdThrowsCorrectExceptionType() {
+		$id = new PropertyId( 'P1' );
+
+		$lookup = new InMemoryEntityLookup();
+		$lookup->addException( new EntityLookupException( $id ) );
+
+		$this->expectException( PropertyLookupException::class );
+		$lookup->getPropertyForId( $id );
 	}
 
 }